Refactoring randomization #2
